Added a new MimeReader
The MimeReader is an alternative to the MimeParser.

Unlike the MimeParser, the MimeReader does not construct a MimeMessage
or MimeEntity "DOM". Instead, it works more like a SAX parser in that
it emits events as logical components of a MIME structure are parsed,
allowing the developer to process MIME data as it is parsed rather than
waiting for the entire message or entity to be completely constructed.

This also allows developers a way to reduce their memory overhead.

Implements the API discussed in issue #695